+++ dumux/assembly/boxlocalassembler.hh
@@ -58,7 +58,6 @@ class BoxLocalAssemblerBase : public FVLocalAssemblerBase<TypeTag, Assembler, Im
     using ParentType = FVLocalAssemblerBase<TypeTag, Assembler, Implementation, implicit>;
     using JacobianMatrix = GetPropType<TypeTag, Properties::JacobianMatrix>;
     using GridVariables = GetPropType<TypeTag, Properties::GridVariables>;
-    using SolutionVector = GetPropType<TypeTag, Properties::SolutionVector>;
     using ElementVolumeVariables = typename GetPropType<TypeTag, Properties::GridVolumeVariables>::LocalView;
 
     enum { numEq = GetPropType<TypeTag, Properties::ModelTraits>::numEq() };
@@ -78,8 +77,8 @@ public:
      * \brief Computes the derivatives with respect to the given element and adds them
      *        to the global matrix. The element residual is written into the right hand side.
      */
-    template <class PartialReassembler = DefaultPartialReassembler>
-    void assembleJacobianAndResidual(JacobianMatrix& jac, SolutionVector& res, GridVariables& gridVariables,
+    template <class ResidualVector, class PartialReassembler = DefaultPartialReassembler>
+    void assembleJacobianAndResidual(JacobianMatrix& jac, ResidualVector& res, GridVariables& gridVariables,
                                      const PartialReassembler* partialReassembler = nullptr)
     {
         this->asImp_().bindLocalViews();
@@ -182,7 +181,8 @@ public:
     /*!
      * \brief Assemble the residual only
      */
-    void assembleResidual(SolutionVector& res)
+    template <class ResidualVector>
+    void assembleResidual(ResidualVector& res)
     {
         this->asImp_().bindLocalViews();
         const auto residual = this->evalLocalResidual();